riverpodで問題状態管理スタート
