+++ lib/main.dart
@@ -1,13 +1,12 @@
-
 import 'package:flutter/material.dart';
-import 'package:writing_learner/screen/home_screen.dart';
-import 'package:writing_learner/screen/question_screen.dart';
-import 'package:writing_learner/screen/question_view.dart';
+import 'package:writing_learner/screens/home_screen.dart';
+import 'package:writing_learner/screens/question_screen.dart';
+import 'package:writing_learner/screens/question_view.dart';
 import 'package:writing_learner/themes/app_theme.dart';
-
+import 'package:flutter_riverpod/flutter_riverpod.dart';
 
 void main() {
-  runApp(const MyApp());
+  runApp(const ProviderScope(child: MyApp()));
 }
 
 class MyApp extends StatelessWidget {
@@ -17,13 +16,12 @@ class MyApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
-      title: 'Writing learner',
-      theme: appTheme(),
-      home: const HomeScreen(),
-      routes:{
-        HomeScreen.routeName: (ctx) => const HomeScreen(),
-        QuestionView.routeName: (ctx) => const QuestionView(),
-      }
-    );
+        title: 'Writing learner',
+        theme: appTheme(),
+        home: const HomeScreen(),
+        routes: {
+          HomeScreen.routeName: (ctx) => const HomeScreen(),
+          QuestionView.routeName: (ctx) => const QuestionView(),
+        });
   }
 }
