+++ lib/screens/question_view.dart
@@ -1,9 +1,7 @@
-import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
-import 'package:writing_learner/screen/question_screen.dart';
-import 'package:writing_learner/screen/question_start_screen.dart';
-import 'package:flutter_dotenv/flutter_dotenv.dart';
-import 'package:google_generative_ai/google_generative_ai.dart';
+import 'package:writing_learner/screens/question_screen.dart';
+import 'package:writing_learner/screens/question_start_screen.dart';
+import 'package:writing_learner/utilities/generative_content.dart';
 
 class QuestionView extends StatefulWidget {
   const QuestionView({super.key});
@@ -13,21 +11,13 @@ class QuestionView extends StatefulWidget {
 }
 
 class QuestionViewState extends State<QuestionView> {
-  List<Widget> questionPages = [QuestionStartScreen()];
+  List<Widget> questionPages = [const QuestionStartScreen()];
   var isInit = true;
   var isLoading = true;
-  Future<String> generateText(String prompt) async {
-    await dotenv.load(fileName: '.env');
-    String apiKey = dotenv.get('GEMINI_API_KEY');
 
-    final model =
-        GenerativeModel(model: 'gemini-1.5-flash-latest', apiKey: apiKey);
-    final content = [Content.text(prompt)];
-    final response = await model.generateContent(content);
-    return response.text!;
-  }
+  var availableQuestionPages = [];
 
-  Future<void> preloadNextPage(int page) async {
+  Future<void> initPages(int page) async {
     /*var japaneseSentences = [
       "近年、環境問題に対する意識が世界中で高まり、再生可能エネルギーの重要性が強調されています。",
       "古代の哲学者たちは、人間の存在意義について多くの議論を交わし、その影響は現代にも及んでいます。",
@@ -37,15 +27,19 @@ class QuestionViewState extends State<QuestionView> {
     ];*/
 
     for (int i = 0; i < 5; i++) {
-      String questionSentence = await generateText(
-          '大学入試対策になるような英訳問題の和文をランダムに出力して。ただし問題の和文のみ一文を出力すること。');
-
+      String questionSentence = await GenerativeService().generateText('大学入試対策になるような英訳問題の和文をランダムに出力して。ただし問題の和文のみ一文を出力すること。');
       questionPages.add(QuestionScreen(question: questionSentence));
     }
-    questionPages.add(QuestionStartScreen());
+    questionPages.add(const QuestionStartScreen());
+    availableQuestionPages.add(questionPages[0]);
+    availableQuestionPages.add(questionPages[1]);
     setState(() {
       isLoading = false;
     });
+  }
+
+  void preloadNextPage(int page) {
+    availableQuestionPages.add(questionPages[page + 1]);
   }
 
   @override
@@ -59,7 +53,7 @@ class QuestionViewState extends State<QuestionView> {
         actions: [],
       ),
       body: isLoading
-          ? Center(
+          ? const Center(
               child: Column(
                 children: [CircularProgressIndicator(), Text('あなたに最適な問題を作成中')],
               ),
