+++ pubspec.yaml
@@ -37,6 +37,7 @@ dependencies:
   cupertino_icons: ^1.0.6
   google_generative_ai: ^0.4.3
   flutter_dotenv: ^5.1.0
+  flutter_riverpod: ^2.5.1
 
 dev_dependencies:
   flutter_test:
